Add fai dependence for bcftools with target fastas
I had some serious bad juju going on here.  bcftools needs the .fai, but it wasn't created anywhere in the workflow.  Not only that, but the fasta used for variant calling was the one in the config, not the version that had been copied to resources.  I think I have fixed that all up now...I hope so!!